requirements.txt
Flask==2.0.2
Flask_Bcrypt==0.7.1
Flask_RESTful==0.3.9
validate-email-address
py3dns
bleach
redis
Flask-APScheduler
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
python-decouple